python/requirements.txt
colorama
dropbox
python-dotenv
tzdata
